+++ 2024/src/01/main.zig
@@ -7,7 +7,7 @@ const part_2 = @import("part_2.zig");
 pub fn main() !void {
     try common.run(
         parse.Input,
-        u32,
+        u64,
         parse.parse_file,
         part_1.run,
         part_2.run,
