+++ 2024/src/01/part_2_test.zig
@@ -3,7 +3,7 @@ const testing = std.testing;
 const parse = @import("parse.zig");
 const part_2 = @import("part_2.zig");
 
-const EXAMPLE_ANSWER_2: u32 = 31;
+const EXAMPLE_ANSWER_2: u64 = 31;
 
 test "part 2 example" {
     var gpa = std.heap.GeneralPurposeAllocator(.{}){};
@@ -16,7 +16,7 @@ test "part 2 example" {
     try testing.expectEqual(result, EXAMPLE_ANSWER_2);
 }
 
-const REAL_ANSWER_2: u32 = 27384707;
+const REAL_ANSWER_2: u64 = 27384707;
 
 test "part 2 real" {
     var gpa = std.heap.GeneralPurposeAllocator(.{}){};
