+++ 2024/src/02/part_2_test.zig
@@ -3,7 +3,7 @@ const testing = std.testing;
 const parse = @import("parse.zig");
 const part_2 = @import("part_2.zig");
 
-const EXAMPLE_ANSWER_2: u32 = 4;
+const EXAMPLE_ANSWER_2: u64 = 4;
 
 test "part 2 example" {
     var gpa = std.heap.GeneralPurposeAllocator(.{}){};
@@ -16,7 +16,7 @@ test "part 2 example" {
     try testing.expectEqual(result, EXAMPLE_ANSWER_2);
 }
 
-const REAL_ANSWER_2: u32 = 364;
+const REAL_ANSWER_2: u64 = 364;
 
 test "part 2 real" {
     var gpa = std.heap.GeneralPurposeAllocator(.{}){};
