+++ 2024/src/06/part_1_test.zig
@@ -3,7 +3,7 @@ const testing = std.testing;
 const parse = @import("parse.zig");
 const part_1 = @import("part_1.zig");
 
-const EXAMPLE_ANSWER_1: u32 = 41;
+const EXAMPLE_ANSWER_1: u64 = 41;
 
 test "part 1 example" {
     var gpa = std.heap.GeneralPurposeAllocator(.{}){};
@@ -16,7 +16,7 @@ test "part 1 example" {
     try testing.expectEqual(EXAMPLE_ANSWER_1, result);
 }
 
-const REAL_ANSWER_1: u32 = 5551;
+const REAL_ANSWER_1: u64 = 5551;
 
 test "part 1 real" {
     var gpa = std.heap.GeneralPurposeAllocator(.{}){};
